Changed WeatherGroup, now using WeatherPhenomena class to store weather info.
Added weatherPhenomena() method to get a WeatherPhenomena type for this group.
The qualifier(), descriptor(), weather(), and contains() methods are still present but deprecated and planned to be removed in version 4.0.0.
The WeatherGroup class is backward compatible with the exception that TS descriptor is allowed with any precipitation and FZ descriptor is allowed with UP precipitation type.