Further tweaked WeatherGroup
Method weatherPhenomena() returns a vector rather than a single WeatherPhenomena value.
Now the planned use of WeatherGroup is to store the information about current, recent or forecast weather phenomena used in main part of METAR or TAF forecast, as well as beginning and ending of the weather phenomena given in the remarks.
WeatherBeginEndGroup is deprecated is planned be removed in version 4.0.0.