LightningGroup and VicinityGroup are now using distance() to indicate distant (DSNT) phenomena
The methods isDistant() are deprecated in both LightningGroup and VicinityGroup.
Distance in form of non-reported value with modifier Distance::Modifier::DISTANT is used in LightningGroup and VicinityGroup to indicate distant phenomena (coded DSNT in the remarks).
Added distance() method to LightningGroup which currently reports only non-reported value or non-reported value with modifier Distance::Modifier::DISTANT (to be used instead of isDistant() when it is removed in version 4.0.0).