Overhaul of CloudTypesGroup, renamed CloudLayersGroup to LowMidHighCloudGroup
Added data type CloudType which can describe cloud type/genus (e.g. Stratocumulus or Altocumulus) or obscuration (e.g. Blowing Snow or Fog) along with base height and precise okta.
Renamed CloudGroup::cloudType() to CloudGroup::convectiveType() for clarity.
Added method CloudGroup::cloudType() which returns a CloudType instance if the cloud layer info is stored in this group or CloudInfo or type 'not reported' if other information is stored in this instance of CloudGroup.
Redesigned CloudGroup public members, cloudTypes() now returns vector of CloudType's stored in the group.
Renamed CloudLayersGroup to LowMidHighCloudGroup for clarity.